# openapi.yaml
openapi: 3.1.0
...
info:
  title: LAGALia — Gestió de contractació pública
  version: 0.1.0
  description: |
    API de la plataforma de gestió de contractació pública de l'Ajuntament de Cunit.

    **API first**: tota funcionalitat del sistema és accessible per aquesta API.
    La interfície web n'és un client més, igual que els agents d'IA, n8n o
    qualsevol integració de tercers.

    Convencions generals a `docs/05-api.md`.
  contact:
    name: Ajuntament de Cunit
  license:
    name: EUPL-1.2
    identifier: EUPL-1.2
servers:
  - url: /api/v1
    description: Servidor actual
...
security:
  - bearerAuth: []
paths:
...
  /settings:
    get:
      tags: [config]
      operationId: listSettings
      summary: Paràmetres (els secrets només is_set)
      x-required-scopes: [config:read]
      responses:
        "200":
          description: Paràmetres
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Setting" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /settings/{key}:
    put:
      tags: [config]
      operationId: putSetting
      summary: Crea o actualitza un paràmetre
      x-required-scopes: [config:write]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string, minLength: 2, maxLength: 100 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value: {}
                description: { type: [string, "null"], maxLength: 500 }
                is_secret: { type: boolean, default: false }
      responses:
        "200":
          description: Paràmetre
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Setting" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /connectors:
    get:
      tags: [config]
      operationId: listConnectors
      summary: Connectors del hub amb estat i salut
      x-required-scopes: [config:read]
      responses:
        "200":
          description: Connectors
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Connector" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /connectors/{slug}:
    patch:
      tags: [config]
      operationId: updateConnector
      summary: Activa/desactiva i configura un connector
      x-required-scopes: [config:write]
      parameters:
        - $ref: "#/components/parameters/ConnectorSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
                config: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Connector
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /connectors/{slug}/credentials:
    put:
      tags: [config]
      operationId: putConnectorCredentials
      summary: Desa credencials (write-only, xifrades)
      x-required-scopes: [config:write]
      parameters:
        - $ref: "#/components/parameters/ConnectorSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credentials]
              properties:
                credentials:
                  type: object
                  additionalProperties: { type: string }
      responses:
        "200":
          description: Connector (credencials com a is_set)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /connectors/smtp/actions/send-test-email:
    post:
      tags: [config]
      operationId: sendSmtpTestEmail
      summary: Envia un correu de prova a l'admin autenticat
      x-required-scopes: [config:write]
      responses:
        "200":
          description: Resultat
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [sent, failed] }
                  detail: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token d'accés d'usuari (JWT curt) o clau de compte de servei (`sk_...`)
        per a agents d'IA i integracions. Sempre a la capçalera `Authorization`;
        mai per query string.
  parameters:
...
    ConnectorSlug:
      name: slug
      in: path
      required: true
      schema: { type: string, minLength: 2, maxLength: 50, pattern: "^[a-z0-9_-]+$" }
...
  responses:
...
    Unauthorized:
      description: Cal autenticar-se o la sessió ha caducat
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Autenticat però sense permís per a aquesta acció o recurs
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Recurs inexistent o fora del teu abast
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationError:
      description: Dades d'entrada no vàlides
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: Error segons RFC 9457. Mai inclou el cos de la petició original.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ["https://lagalia/errors/validation"] }
        title: { type: string, examples: ["Dades no vàlides"] }
        status: { type: integer, examples: [422] }
        detail: { type: string }
        instance: { type: string }
        trace_id:
          type: string
          description: Identificador de traça per a suport.
        errors:
          type: array
          description: Errors per camp, quan aplica.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    Setting:
      type: object
      required: [key, is_secret, is_set]
      properties:
        key: { type: string }
        value: {}
        description: { type: [string, "null"] }
        is_secret: { type: boolean }
        is_set: { type: boolean }
        placeholder: { type: string }
    Connector:
      type: object
      required: [slug, name, enabled, mode, config, config_defaults, credentials]
      properties:
        slug: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        mode: { type: string }
        config: { type: object, additionalProperties: true }
        config_defaults: { type: object, additionalProperties: true }
        credentials:
          type: object
          additionalProperties: { type: boolean }
          description: "nom de credencial → is_set (mai el valor)."
        health_status: { type: [string, "null"] }
        last_health_check: { type: [string, "null"], format: date-time }
...